Converted and moved the getSharedLayers function to react hooks.
